+++ src/argon2/_password_hasher.py
@@ -145,20 +145,34 @@ class PasswordHasher:
     def type(self) -> Type:
         return self._parameters.type
 
-    def hash(self, password: str | bytes) -> str:
+    def hash(self, password: str | bytes, *, salt: bytes | None = None) -> str:
         """
         Hash *password* and return an encoded hash.
 
-        :param password: Password to hash.
-        :type password: ``bytes`` or ``str``
+        Parameters:
 
-        :raises argon2.exceptions.HashingError: If hashing fails.
+            password: Password to hash.
 
-        :rtype: str
+            salt: If None, a random salt is securely created.
+
+                .. danger::
+
+                    You should **not** pass a salt unless you really know what
+                    you are doing.
+
+        Raises:
+
+            argon2.exceptions.HashingError: If hashing fails.
+
+        Returns:
+
+            Hashed *password*.
+
+        .. versionadded:: 23.1.0 *salt* parameter
         """
         return hash_secret(
             secret=_ensure_bytes(password, self.encoding),
-            salt=os.urandom(self.salt_len),
+            salt=salt or os.urandom(self.salt_len),
             time_cost=self.time_cost,
             memory_cost=self.memory_cost,
             parallelism=self.parallelism,